Fix coding suggestions
